DPC-203: Move encrypting aggregation engine test to their own class
